Deduplicate code
And almost handle left-side graphs
