+++ src/main/resources/assets/angelica/shaders/debug_graph.frag.glsl
@@ -4,11 +4,13 @@ uniform int headIdx;
 uniform float scaleFactor;
 uniform float samples[240];
 uniform float pxPerNs;
+uniform bool left;
+uniform float fbWidth;
 
 void main() {
     // Get position - gl_FragCoord starts from the lower left and returns the position of the center of the pixel
-    // i.e. the lower-left-most pixel is (0.5, 0.5)
-    int dx = int(gl_FragCoord.x / scaleFactor);
+    // i.e. the lower-left-most pixel is (0.5, 0.5). Shift if we're on the right side of the screen.
+    int dx = int((left ? gl_FragCoord.x : fbWidth - gl_FragCoord.x) / scaleFactor) - 1;
     int dy = int(gl_FragCoord.y / scaleFactor);
 
     // Get the time for this frag. Last at the left edge, first at the right, shifting as needed when the head moves.
